+++ .idea/encodings.xml
@@ -4,6 +4,9 @@
     <file url="file://$PROJECT_DIR$" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/spark-quickbooks" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/spark-quickbooks-api" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/spark-quickbooks-api/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/spark-quickbooks-runtime" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/spark-quickbooks-runtime/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/spark-quickbooks/src/main/java" charset="UTF-8" />
   </component>
 </project>
