+++ README.md
@@ -15,6 +15,9 @@ If the account is set up correctly you will be able to generate tokens and make 
 Note that the authorization code is one time use so if you need new tokens you will need to first
 get a new authorization code.
 
+If you have already generated an access token you can pass that as an option instead of having
+the data source manage the OAuth flow.
+
 The data source is implemented as `DataSourceV2` which cannot be accessed through the SQL syntax
 in Spark 2.3.2. You will need to use the Java or Scala API to use this connector.
 
@@ -37,30 +40,40 @@ artifactId: spark-quickbooks-api
 version: 1.1.0
 ```
 
-## Options
+## General Options
 
 | Option            | Description                       |
 | ----------------- |---------------------------------- |
-| authorizationCode | OAuth Authorization Code          |
 | companyId         | QuickBooks Online Company ID      |
-| clientId          | OAuth Client ID                   |
-| clientSecret      | OAuthClientSecret                 |
-| redirectUri       | HTTPS endpoint for OAuth redirect |
 | entity            | Object to query from the API      |
 | production        | Query production environment      |
 | expandArrays      | Expands nested arrays to columns  |
 
-* `authorizationCode`: Exchanged for access/refresh tokens
 * `companyId`: Also called `realmId`, it's the ID of the company that you want to query in QuickBooks
-* `clientId`: From your app's **Keys** tab
-* `clientSecret`: From your app's **Keys** tab
-* `redirectUri`: Also in the **Keys** tab, this need to be HTTPS in production and cannot be localhost
-  * Default: `https://developer.intuit.com/v2/OAuth2Playground/RedirectUrl`
 * `entity`: Due to the nature of the QuickBooks Online query syntax, only 1 entity may be queried at a time.
 * `production`: set to `true` when switching from a sandbox to production environment
 * `expandArrays`: `true` to expand every element in an array to its own column
     * `lineItems: [{price: 7.0}, {price: 3.0}]` becomes `lineItems_0_price, lineItems_1_price`
     with the value 7.0 and 3.0 respectively
+
+## OAuth Options
+
+| Option            | Description                       |
+| ----------------- |---------------------------------- |
+| accessToken       | OAuth Access Token                |
+| authorizationCode | OAuth Authorization Code          |
+| clientId          | OAuth Client ID                   |
+| clientSecret      | OAuthClientSecret                 |
+| redirectUri       | HTTPS endpoint for OAuth redirect |
+
+* `accessToken`: OAuth access token if you have already generated one. If you pass an access token
+you don't need to pass any other OAuth options and we'll try to use your token instead of going
+through the authorization process.
+* `authorizationCode`: Exchanged for access/refresh tokens
+* `clientId`: From your app's **Keys** tab
+* `clientSecret`: From your app's **Keys** tab
+* `redirectUri`: Also in the **Keys** tab, this need to be HTTPS in production and cannot be localhost
+  * Default: `https://developer.intuit.com/v2/OAuth2Playground/RedirectUrl`
 
 We take this entity and pass it as the query `select * from <entity>` and then create a data frame
 from the result set to query against with Spark SQL
