+++ spark-quickbooks-runtime/src/main/java/inetsoft/spark/quickbooks/QuickbooksRuntime.java
@@ -15,22 +15,13 @@
  */
 package inetsoft.spark.quickbooks;
 
-import com.intuit.ipp.core.*;
 import com.intuit.ipp.exception.FMSException;
-import com.intuit.ipp.security.OAuth2Authorizer;
-import com.intuit.ipp.services.*;
-import com.intuit.ipp.util.Config;
+import com.intuit.ipp.exception.InvalidTokenException;
+import com.intuit.ipp.services.QueryResult;
 import com.intuit.oauth2.client.OAuth2PlatformClient;
-import com.intuit.oauth2.config.Environment;
-import com.intuit.oauth2.config.OAuth2Config;
-import com.intuit.oauth2.data.BearerTokenResponse;
 import com.intuit.oauth2.exception.OAuthException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.lang.invoke.MethodHandles;
-import java.util.ArrayList;
-import java.util.List;
+import inetsoft.spark.quickbooks.token.TokenStrategy;
+import inetsoft.spark.quickbooks.token.TokenStrategyFactory;
 
 public class QuickbooksRuntime implements QuickbooksAPI {
    public QuickbooksQueryResult loadData(String accessToken,
@@ -42,42 +33,17 @@ public class QuickbooksRuntime implements QuickbooksAPI {
                                          boolean production,
                                          String entity)
    {
-      this.clientId = clientId;
-      this.clientSecret = clientSecret;
-      this.authorizationCode = authorizationCode;
-      this.companyId = companyId;
-      this.redirectUrl = redirectUrl;
-      this.production = production;
-      this.entity = entity;
-
-      if(accessToken == null || accessToken.isEmpty()) {
-         return loadData(QuickbooksConfig.readConfig(clientId, companyId), null);
-      }
-
-      return loadData(null, accessToken);
-   }
-
-   /**
-    * Call the Quickbooks API to check the tokens and execute our query
-    */
-   private QuickbooksQueryResult loadData(QuickbooksConfig config, String accessToken) {
       final ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
 
       try {
          Thread.currentThread().setContextClassLoader(getClass().getClassLoader());
-
-         // check tokens
-         OAuth2Config oauth2Config = new OAuth2Config.OAuth2ConfigBuilder(clientId, clientSecret)
-            .callDiscoveryAPI(production ? Environment.PRODUCTION : Environment.SANDBOX)
-            .buildConfig();
-         client = new OAuth2PlatformClient(oauth2Config);
-
-         if(accessToken == null) {
-            accessToken = connect(config);
-         }
-
-         final QueryResult queryResult = execute(accessToken);
-         return new QueryResultAdapter(queryResult);
+         final TokenStrategy tokenStrategy = TokenStrategyFactory.create(accessToken, clientId,
+                                                                         clientSecret, companyId,
+                                                                         authorizationCode,
+                                                                         production, redirectUrl);
+         final String token = tokenStrategy.getAccessToken();
+         final QueryResult result = queryExecutor.execute(token, companyId, production, entity);
+         return new QueryResultAdapter(result);
       }
       catch(OAuthException e) {
          throw new RuntimeException("OAuth authentication failed", e);
@@ -91,147 +57,5 @@ public class QuickbooksRuntime implements QuickbooksAPI {
       }
    }
 
-   private QueryResult execute(String accessToken) throws FMSException {
-      final String apiUrl = production ? productionUrl : sandboxUrl;
-      Config.setProperty(Config.BASE_URL_QBO, apiUrl);
-      final OAuth2Authorizer oauth = new OAuth2Authorizer(accessToken);
-      final Context context = new Context(oauth, ServiceType.QBO, companyId);
-      final DataService service = new DataService(context);
-
-      // first execute a count query to determine pagination
-      final int totalCount = getTotalCount(service);
-
-      // next execute paginated results until complete
-      final QueryResult queryResult = new QueryResult();
-      int startPosition = 1;
-      queryResult.setStartPosition(startPosition);
-      queryResult.setTotalCount(totalCount);
-      queryResult.setMaxResults(totalCount);
-      final ArrayList<IEntity> entities = new ArrayList<>();
-      BatchOperation batchOperation = new BatchOperation();
-      int counter = 0;
-
-      for(int remaining = totalCount; remaining > 0; remaining -= RESULT_LIMIT) {
-         final int maxResults = Math.min(RESULT_LIMIT, remaining);
-         final String query = String.format("SELECT * FROM %s STARTPOSITION %d MAXRESULTS %d",
-                                            entity,
-                                            startPosition,
-                                            maxResults);
-         batchOperation.addQuery(query, String.valueOf(counter++));
-         startPosition += RESULT_LIMIT;
-
-         if(counter % BATCH_LIMIT == 0) {
-            executeBatchOperation(service, entities, batchOperation);
-            batchOperation = new BatchOperation();
-         }
-      }
-
-      executeBatchOperation(service, entities, batchOperation);
-      queryResult.setEntities(entities);
-      return queryResult;
-   }
-
-   private void executeBatchOperation(DataService service,
-                                      ArrayList<IEntity> entities,
-                                      BatchOperation batchOperation) throws FMSException
-   {
-      final List<String> bIds = batchOperation.getBIds();
-
-      if(bIds.size() > 0) {
-         final String index = bIds.get(0);
-         LOG.debug("Executing QuickBooks query from index: {}", index);
-         service.executeBatch(batchOperation);
-
-         for(String bId : bIds) {
-            final QueryResult queryResponse = batchOperation.getQueryResponse(bId);
-            entities.addAll(queryResponse.getEntities());
-         }
-      }
-   }
-
-   /**
-    * Get the total number of entities in the query response
-    */
-   private int getTotalCount(DataService service) throws FMSException {
-      final QueryResult countResult = service.executeQuery("SELECT COUNT(*) FROM " + entity);
-      final Integer totalCount = countResult.getTotalCount();
-      LOG.debug("QuickBooks count returned {} result(s)", totalCount);
-      return totalCount != null ? totalCount : 1;
-   }
-
-   /**
-    * Handle OAuth connection. If the access token is null retrieve new tokens. If the refresh token
-    * has expired and the access token is not null refresh the current access token. Otherwise don't
-    * call any authorization mechanism
-    *
-    * @return the current and valid access token
-    */
-   private String connect(QuickbooksConfig config) throws OAuthException {
-      final String accessToken = config.getAccessToken();
-      final long expiration = config.getExpiration();
-
-      if(accessToken == null) {
-         LOG.debug("Fetching OAuth tokens");
-         final BearerTokenResponse response =
-            client.retrieveBearerTokens(authorizationCode, redirectUrl);
-         config = config.updateCredentials(response.getExpiresIn(),
-                                           response.getAccessToken(),
-                                           response.getRefreshToken(), clientId, companyId);
-      }
-      else {
-         if(expiration > -1 && expiration < System.currentTimeMillis()) {
-            LOG.debug("Refreshing OAuth Tokens");
-            final BearerTokenResponse response;
-            response = client.refreshToken(config.getRefreshToken());
-            config = config.updateCredentials(response.getExpiresIn(),
-                                              response.getAccessToken(),
-                                              response.getRefreshToken(), clientId, companyId);
-         }
-      }
-
-      return config.getAccessToken();
-   }
-
-   public static class QueryResultAdapter implements QuickbooksQueryResult {
-      public QueryResultAdapter(QueryResult queryResult) {
-         this.queryResult = queryResult;
-      }
-
-      @Override
-      public List<Object> getEntities() {
-         return new ArrayList<>(queryResult.getEntities());
-      }
-
-      @Override
-      public int getStartPosition() {
-         return queryResult.getStartPosition();
-      }
-
-      @Override
-      public int getMaxResults() {
-         return queryResult.getMaxResults();
-      }
-
-      @Override
-      public int getTotalCount() {
-         return queryResult.getTotalCount();
-      }
-
-      private QueryResult queryResult;
-   }
-   // max 30 queries per batch operation
-   public static final int BATCH_LIMIT = 30;
-   private static final String sandboxUrl = "https://sandbox-quickbooks.api.intuit.com/v3/company";
-   private static final String productionUrl = "https://quickbooks.api.intuit.com/v3/company";
-   private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-   // max number of results quickbooks can return
-   private static final int RESULT_LIMIT = 1000;
-   private OAuth2PlatformClient client;
-   private String clientId;
-   private String clientSecret;
-   private String authorizationCode;
-   private String companyId;
-   private String redirectUrl;
-   private boolean production;
-   private String entity;
+   private final QueryExecutor queryExecutor = new QueryExecutorService();
 }
