+++ spark-quickbooks-runtime/src/main/java/inetsoft/spark/quickbooks/token/ExternalOAuthTokenStrategy.java
@@ -1,0 +1,18 @@
+package inetsoft.spark.quickbooks.token;
+
+/**
+ * This strategy assumes that OAuth is handled outside of the data source so we just use the given
+ * access token
+ */
+public class ExternalOAuthTokenStrategy implements TokenStrategy {
+   public ExternalOAuthTokenStrategy(String accessToken) {
+      this.accessToken = accessToken;
+   }
+
+   @Override
+   public String getAccessToken() {
+      return accessToken;
+   }
+
+   private final String accessToken;
+}
