+++ spark-quickbooks/src/main/java/inetsoft/spark/quickbooks/source/DefaultSource.java
@@ -27,11 +27,11 @@ public class DefaultSource implements DataSourceV2, ReadSupport, DataSourceRegis
 
    @Override
    public DataSourceReader createReader(DataSourceOptions options) {
-      final String clientId = options.get("clientId").orElse("");
-      final String clientSecret = options.get("clientSecret").orElse("");
-      final String authorizationCode = options.get("authorizationCode").orElse("");
-      final String accessToken = options.get("accessToken").orElse("");
-      final String companyId = options.get("companyId").orElse("");
+      final String clientId = options.get("clientId").orElse(null);
+      final String clientSecret = options.get("clientSecret").orElse(null);
+      final String authorizationCode = options.get("authorizationCode").orElse(null);
+      final String accessToken = options.get("accessToken").orElse(null);
+      final String companyId = options.get("companyId").orElse(null);
       final String redirectUrl = options.get("redirectUrl").orElse(URL);
       final String entity = options.get("entity").orElse("companyInfo");
       final boolean expandArrays = options.get("expandArrays")
